+++ Engine/src/main/java/com/ing/engine/commands/webservice/Webservice.java
@@ -28,7 +28,7 @@ import org.xml.sax.InputSource;
 import java.io.StringReader;
 import java.net.URLEncoder;
 import java.net.http.HttpClient;
-import java.net.http.HttpClient.Redirect;
+
 import java.net.http.HttpHeaders;
 import java.net.http.HttpRequest;
 import java.net.http.HttpRequest.BodyPublisher;
@@ -1169,36 +1169,6 @@ public class Webservice extends GeneralWebservice {
         }
     }
 
-    /**
-     * Retrieves the HTTP redirect policy configured for the current API driver settings.
-     * <p>
-     * The logic follows three strict rules:
-     * <ul>
-     *   <li>If no value is configured (i.e., the property is {@code null} or blank), the method defaults to
-     *       {@link Redirect#NEVER}.</li>
-     *   <li>If a valid redirect policy is provided (one of {@code NEVER}, {@code NORMAL}, or {@code ALWAYS},
-     *       case-insensitive), the corresponding {@link Redirect} enum is returned.</li>
-     *   <li>If a value is provided but does not match any {@link Redirect} enum constant, the method throws an
-     *       {@link IllegalArgumentException} to indicate a configuration error.</li>
-     * </ul>
-     * </p>
-     *
-     * @return the resolved {@link Redirect} policy to be applied when building the {@link java.net.http.HttpClient}
-     * @throws IllegalArgumentException if a non-blank but invalid redirect value is configured
-     */
-    private Redirect getRedirectPolicy() {
-        String httpClientRedirect = Control.getCurrentProject().getProjectSettings().getDriverSettings().getHttpClientRedirect();
-
-        if (httpClientRedirect == null || httpClientRedirect.trim().isEmpty()) {
-            return Redirect.NEVER;
-        }
-
-        try {
-            return Redirect.valueOf(httpClientRedirect.trim().toUpperCase());
-        } catch (IllegalArgumentException ex) {
-            throw new IllegalArgumentException("Invalid httpClientRedirect value: '" + httpClientRedirect + "'. Allowed values: NEVER, NORMAL, ALWAYS.");
-        }
-    }
 
     /**
      * Extracts a cookie value from the HTTP response headers and stores it in a variable.
